Add full KEPN data.
This is from kepn.nottingham.ac.uk/places/lang/n/geojson, which is
already filtered to just Norse names, but is in a slightly different
format to what we were using before. I've mapped it down to the same
format, and then removed extra elements that we weren't using, as well
as doing the sorting by Norse eytomological elements again.