Re-organise the kepn data to be indexed by placename.
There are multiple places with identical names, eg: Thorpe, but to
make matching geocoding results to a Norse place easy and fast, it
would be nice to be able to look them up. The solution here is to have
an object which matches name -> list of places with that name.

The worst case is 5 (for Thorpe), which is a nice small list that we
can then loop over and compare distances to get a better match.