Add functions to find Norse places for geocode results.
This commit adds functions to find the slug of a norse name from geocode results
as well as re-arranging the functions which build the HTML for said results so
that the slug is passed on in links. We also add code to deal with clicks on
results that have a corresponding Norse place differently, and a new index of
the map markers, by slug, so that they can be looked up efficiently.