Check for markers when showing a point and alert if there are none
This is a bit of a brute force way of doing things, but there's no
better way. The code simply loops over every marker once we've zoomed
to a specific point and checks if it's in the map bounds. If there are
no markers in the map bounds then it alerts the user.

This is set to happen whenever we zoom into a place as a result of
a search for a place that didn't return a Norse place already, or
when someone uses their geolocation, not when we zoom in normally
because presumably then you can see their are no places and don't need
it explaining to you.